refactor: centralize environment detection and normalization
- Create `constants/environment.ts` to manage `ENV` and `normalizeEnvironment` logic.
- Replace repetitive environment switch-case logic with `normalizeEnvironment` in `endpoint-config.ts` and `environment-variables.ts`.
- Update multiple files to import `ENV` from the new `constants/environment` module.
- Add JSDoc headers to `config.ts`, `endpoint-config.ts`, and `environment-variables.ts` to clarify their purpose (Build-time vs. Runtime).
