+++ packages/mirror-tv-next/app/robots.txt/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from 'next/server'
-import { ENV } from '~/constants/environment-variables'
+import { ENV } from '~/constants/environment'
 
 export async function GET() {
   let robotsContent = ''
