+++ packages/mirror-tv-next/constants/config.ts
@@ -1,3 +1,10 @@
+/**
+ * Runtime Config
+ *
+ * Server-side process.env values without NEXT_PUBLIC_ prefix.
+ * Cloud Run can override these at deploy time without rebuilding the image.
+ */
+
 const API_ENDPOINT_OVERRIDE_FROM_ENV = process.env.API_ENDPOINT
 const ENABLE_K6_NEW_ENDPOINTS = process.env.ENABLE_K6_NEW_ENDPOINTS === 'true'
 const PROGRAMMABLE_SEARCH_API_KEY = process.env.PROGRAMMABLE_SEARCH_API_KEY
